+++ uno/Draw_Stop_Display.py
@@ -1,14 +1,14 @@
-from Button_Class import Button
-from Text_Class import Text
+from uno.Button_Class import Button
+from uno.Text_Class import Text
 import pygame as pg
 
 Button_list = []
-Button_list.append(Button((400, 200), (100, 50), 'SINGLE PLAYER'))
+Button_list.append(Button((400, 200), (100, 50), 'RESUME'))
 Button_list.append(Button((400, 300), (100, 50), 'OPTIONS'))
-Button_list.append(Button((400, 400), (100, 50), 'QUIT'))
+Button_list.append(Button((400, 400), (100, 50), 'GAME QUIT'))
 
 Text_list = []
-Text_list.append(Text((320, 60), 40, 'UNO Game'))
+Text_list.append(Text((320, 60), 40, 'PAUSE'))
 
 def draw(screen, mode, running):
     for event in pg.event.get():
@@ -24,13 +24,13 @@ def draw(screen, mode, running):
             for idx, item in enumerate(Button_list):
                 if item.rect.collidepoint(mouse_pos):
                     item.active()
-                    if idx==0:                            
+                    if idx==0:
                         mode[0]=3
                     elif idx==1:
                         mode[0]=2
-                        mode[1]=1
+                        mode[1]=4
                     elif idx==2:
-                        running[0] = False
+                        mode[0]=1
         else:
             if flag:
                 for item in Button_list:
