+++ uno/Text_Class.py
@@ -1,5 +1,5 @@
 import pygame as pg
-from Color import *
+from uno.Color import *
 
 class Text:
     def __init__(self, pos, size, text, color = BLACK): # pos = 좌상단 모서리 좌표, size = 글씨 크기, text = 내용
